Use ROUTER_LOGDIR in mysqlrouter logrotate config
In AUR MySQL [1], ROUTER_INSTALL_LOGDIR is set to var/log/mysqlrouter.
Since this is a relative path, according to [2], it should be prefixed
with CMAKE_INSTALL_PREFIX.

However in the final package, var/log/mysqlrouter was used directly and
this caused the error below:

> logrotate[385400]: mysqlrouter:31 keyword 'var' not properly separated, found 0x2f
> logrotate[385400]: error: mysqlrouter:31 unknown option 'var' -- ignoring line
> logrotate[385400]: error: mysqlrouter:45 unexpected }
> logrotate[385400]: error: found error in file mysqlrouter, skipping

So I think ROUTER_LOGDIR should be used instead in this template.

[1] https://aur.archlinux.org/packages/mysql
[2] https://github.com/mysql/mysql-server/blob/6846e6b2f72931991cc9fd589dc9946ea2ab58c9/router/cmake/configure.cmake#L50-L62